fix when withdrawal > balance

+++ menu.h
@@ -2,5 +2,5 @@
 #include<map>
 using namespace std;
 int guiTien(map<int, int> &soTo, int loai[]);
-int rutTien(map<int, int> &soTo, int loai[], int balance);
+int rutTien(map<int, int>& soTo, int loai[], unsigned long long int balance);
 void vanTin(const long long int &balance);
